+++ R/check_code_style.R
@@ -1,8 +1,5 @@
-check_code_style = function() {
-  language = get_repo_language()
-  if (language == "python") return(invisible(NULL))
-
-    message(yellow(symbol$circle_filled, "Checking lint...check_code_style()"))
+check_r_style = function() {
+  message(yellow(symbol$circle_filled, "Checking lint...check_code_style()"))
   if (isFALSE(config::get("lintr"))) {
     return(invisible(NULL))
   }
@@ -26,6 +23,37 @@ check_code_style = function() {
       message(red(msg))
       bad_lints = TRUE
     }
+  }
+  return(bad_lints)
+}
+
+
+check_python_style = function() {
+  message(yellow(symbol$circle_filled, "Checking lint...check_code_style()"))
+  if (!fs::file_exists("flake8.ini")) {
+    message("Missing ", "flake8_config_Rmd.ini" ," file - creating a default")
+    flake8_ini_sys = system.file("","flake8_config_Rmd.ini", package = "jrNotes", mustWork = TRUE)
+    file.copy(flake8_ini_sys, to = "flake8_config.ini", overwrite = TRUE)
+  }
+  message(yellow(symbol$circle_filled, "Checking lint...check_code_style()"))
+  fnames = list.files(pattern = "^c.*Rmd$")
+  bad_lints = FALSE
+  for (i in seq_along(fnames)) {
+    message("  ", yellow(symbol$circle_filled, "Checking ", fnames[i]))
+    jrpytests = reticulate::import("jrpytests")
+    reticulate::jrpytests$runflake8rmdpychunks(filename = fnames[i])
+
+  }
+  return(bad_lints)
+
+}
+
+check_code_style = function() {
+  language = get_repo_language()
+  if (language == "python") {
+    bad_lints = check_python_style()
+  } else {
+    bad_lints = check_r_style()
   }
 
   if (bad_lints) {