+++ NEWS.md
@@ -1,5 +1,5 @@
 # Dev
-
+  * Adding ggplot2 to suggests - required for some notes
 # 0.5.1
 ## Cleaners