+++ NEWS.md
@@ -1,4 +1,5 @@
 # Dev
+  * Update missing packages and clean
 
 # 0.5.1
 ## Cleaners